+++ CHANGES.rst
@@ -12,6 +12,8 @@ Changes
   ``writers`` module does not belong to the public API anyway.
 * Dropped testing against PyPy2
 * Changed default Python test version to 3.11 (2.7 is still supported)
+* Support for Python 2.7 will be dropped because it is no longer possible to run automatic
+  tests against the version.
 
 
 1.5.2 -- 2022-05-25